+++ CHANGELOG.md
@@ -50,6 +50,8 @@
   `Plutarch.Internal.Ord`
 * `pif`, `pif'`, `pand`, `pand'`, `por`, `por'`, `pnot`, `#&&`, `#||` are 
    now in `Plutarch.Builtin.Bool`
+* `Term s PRational` is now `Fractional` directly, instead of by way of
+  `PFractional`
 
 ## Removed
 
@@ -63,6 +65,7 @@
 * `Plutarch` module, as it served no useful purpose and was just confusing
 * `PType` synonym (use `S -> Type` honestly instead)
 * `PSBool` and functionality (now in `plutarch-ledger-api`)
+* `PFractional` type class (only one instance, unlikely to ever have more)
 
 ### Fixed